Version 1.0
Currently the script does the following:

run the lista_pastas function, which lists all files, folders and subfolders on the root_folder and then insert this info into a dictionary where the key is the folder path and the values are the files

Run the sobe_pastas() function, which reads the file created by the function lista_pastas() and then create each folder on the GDRIVE parent folder, and gets the ID of the created folder and add it with the path to another dictionary called pasta_anterior. After creating all folder and subfolders the function uploads each file using both the files_and_folders dictionary created.